move some files to components
